pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jutils"
dynamic = ["version"]                     # <- version comes from jutils.__version__
description = "Collection of useful utility functions."
readme = { file = "README.md", content-type = "text/markdown" }
license = { text = "MIT License" }
authors = [{ name = "Johannes Schusterbauer" }]
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0",
    "numpy",
    "einops",
    "pillow",
    "ipython",
    "termcolor",
    "matplotlib",
    "jaxtyping",
]
classifiers = [
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]

[project.urls]
Homepage = "https://github.com/joh-schb/jutils"

[tool.setuptools.dynamic]
version = { attr = "jutils.__version__" }

[tool.setuptools.packages.find]
include = ["jutils*"]